+++ python/lsst/ts/standardscripts/auxtel/daytime_checkout/atpneumatics_checkout.py
@@ -92,7 +92,7 @@ class ATPneumaticsCheckout(salobj.BaseScript):
     def set_metadata(self, metadata):
         """Set estimated duration of the script."""
 
-        metadata.duration = 5.0 * 60  # Approximate 5min to completion
+        metadata.duration = 70
 
     async def run(self):